Added small/medium/large breakpoints with 3/4-column grid sizing
Added 3 media queries with breakpoints to define different column widths with different percentages to introduce the benefits of multiple class names including "small-width-x" and "large-width-x" with a specific naming convention. Created a second page to display how easy it is to use this css naming convention to set up subsequent pages.